requirements.txt
fastmcp>=0.1.0
paho-mqtt>=1.6.1
